+++ libraries/bio/src/tests/monomer-lib-tests.ts
@@ -7,7 +7,7 @@ import {expectObject} from '@datagrok-libraries/utils/src/test';
 import {IMonomerLib, MonomerLibSummaryType} from '../types/index';
 
 /** Summary for HELMCoreLibrary.json */
-export const defaultMonomerLibSummary: MonomerLibSummaryType = {'PEPTIDE': 322, 'RNA': 383, 'CHEM': 0};
+export const defaultMonomerLibSummary: MonomerLibSummaryType = {'PEPTIDE': 324, 'RNA': 383, 'CHEM': 0};
 
 export function expectMonomerLib(lib: IMonomerLib, summary?: { [polymerType: string]: number }): void {
   const exp: MonomerLibSummaryType = summary ?? defaultMonomerLibSummary;